feat(CDF-3933): Add Vercel support to discord bot implementation
