+++ quickstart-template/discord_tutorial/.gitignore
@@ -1,0 +1,1 @@
+.vercel
