+++ quickstart-template/discord_tutorial/vercel.json
@@ -1,0 +1,16 @@
+{
+  "version": 2,
+  "builds": [
+    {
+      "src": "app.cjs",
+      "use": "@now/node"
+    }
+  ],
+  "routes": [
+
+    {
+      "src": "/(.*)",
+      "dest": "app.cjs"
+    }
+  ]
+}
